pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yolo-dataset-analyzer"
version = "1.0.0"
description = "YOLOv9向け初心者フレンドリーなデータセット分析ツール"
authors = [{name = "Claude Code", email = "code@claude.ai"}]
requires-python = ">=3.8"
dependencies = [
    "streamlit>=1.28.0",
    "pandas>=2.0.0",
    "pillow>=10.0.0",
    "plotly>=5.15.0",
    "numpy>=1.24.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.7.0",
    "flake8>=6.0.0"
]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"